examples/finlex-ingestion/requirements.txt
requests==2.31.0
azure-storage-blob==12.19.0
azure-identity==1.15.0
tiktoken==0.6.0
openai==1.12.0
azure-search-documents==11.4.0
